speech-to-text-server/requirements.txt
google-cloud-speech==0.30.0
paho-mqtt==1.4.0
pyaudio
numpy
